+++ continual/transformer.py
@@ -7,7 +7,6 @@ import torch
 from torch import Tensor, nn
 
 from continual.delay import State as DelayState
-from continual.module import TensorPlaceholder
 
 from .closure import Lambda, Unity
 from .container import BroadcastReduce, Residual, Sequential
@@ -38,7 +37,7 @@ class MhaType(Enum):
 class SelectOrDelay(Delay):
     """Select a temporal index during forward, or delay correspondingly during forward_step(s)"""
 
-    def forward(self, x: Tensor) -> Union[Tensor, TensorPlaceholder]:
+    def forward(self, x: Tensor) -> Optional[Tensor]:
         assert len(x.shape) >= 3  # N, C, T
         return x[:, :, -1 - self._delay].unsqueeze(2)
 
@@ -92,6 +91,7 @@ class RetroactiveUnity(Delay):
             new_index = new_index % self.delay
 
         # Get output
+        output = None
         if index >= 0:
             output = buffer.clone().roll(shifts=-index - 1, dims=0)
             idx = (
@@ -102,8 +102,6 @@ class RetroactiveUnity(Delay):
             output = output.permute(
                 list(range(1, idx + 1)) + [0] + list(range(idx + 1, len(output.shape)))
             )
-        else:
-            output = TensorPlaceholder(buffer[0].shape)
 
         return output, (buffer, new_index)